feat: Refactor UniversalSearchDappItem to use getSearchInput and streamline website handling logic
