+++ packages/kit/src/views/UniversalSearch/components/SearchResultItems/UniversalSearchDappItem.tsx
@@ -1,23 +1,40 @@
 import { Image } from '@onekeyhq/components';
 import { ListItem } from '@onekeyhq/kit/src/components/ListItem';
 import { isGoogleSearchItem } from '@onekeyhq/shared/src/consts/discovery';
+import { EEnterMethod } from '@onekeyhq/shared/src/logger/scopes/discovery/scenes/dapp';
 import type { IUniversalSearchDapp } from '@onekeyhq/shared/types/search';
+
+import { useWebSiteHandler } from '../../../Discovery/hooks/useWebSiteHandler';
 
 interface IUniversalSearchDappItemProps {
   item: IUniversalSearchDapp;
-  onPress: () => void;
+  getSearchInput: () => string;
 }
 
 export function UniversalSearchDappItem({
   item,
-  onPress,
+  getSearchInput,
 }: IUniversalSearchDappItemProps) {
   const { name, dappId, logo } = item.payload;
   const isGoogle = isGoogleSearchItem(dappId);
+  const handleWebSite = useWebSiteHandler();
 
   return (
     <ListItem
-      onPress={onPress}
+      onPress={() => {
+        console.log('[universalSearch] renderItem: ', item);
+        handleWebSite({
+          dApp: isGoogle ? undefined : item.payload,
+          // @ts-expect-error
+          webSite: isGoogle
+            ? {
+                title: 'Google',
+                url: getSearchInput(),
+              }
+            : undefined,
+          enterMethod: EEnterMethod.search,
+        });
+      }}
       renderAvatar={<Image source={{ uri: logo }} size="$10" />}
       title={name}
       titleProps={{
